Force package unzip&rebuild before signature overwrite
This is probably overkill and needs a proper method,
but it is needed to avoid a big issue in WAPT custom_zip
implementation.

When a zipfile subdir is listed in custom_zip.filelist,
it will also be signed. And waptclient is not able to check
signatures on directories, leading to a SHA corruption message.

Unziping and rebuilding the packages makes sure there is no bogus
subdir in filelist.